@@ -13,18 +13,18 @@
   "devDependencies": {
     "@playwright/test": "^1.40.1",
     "@sveltejs/adapter-static": "^3.0.1",
-    "@sveltejs/kit": "^2.4.1",
+    "@sveltejs/kit": "^2.5.27",
     "@types/node": "^20.10.6",
     "autoprefixer": "^10.4.14",
     "prettier": "^3.1.1",
-    "prettier-plugin-svelte": "^3.1.2",
-    "svelte": "^4.2.9",
+    "prettier-plugin-svelte": "^3.2.6",
+    "svelte": "^5.0.0",
     "svelte-body": "^1.4.0",
-    "svelte-check": "^3.6.0",
-    "svelte-preprocess": "^5.1.3",
+    "svelte-check": "^4.0.0",
+    "svelte-preprocess": "^6.0.0",
     "tslib": "^2.4.1",
-    "typescript": "^5.0.0",
-    "vite": "^5.0.12"
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4"
   },
   "type": "module",
   "trustedDependencies": [
