@@ -3,16 +3,16 @@
 
   const dispatch = createEventDispatcher();
 
-  let elem: HTMLElement;
+  let elem: HTMLElement = $state();
 
-  let pos: { x: number; y: number } = { x: 0, y: 0 };
+  let pos: { x: number; y: number } = $state({ x: 0, y: 0 });
   let vel: { x: number; y: number } = {
     x: (Math.random() - 0.5) * 5,
     y: Math.random() * -10 - 10,
   };
   let rotationSpeed: number = Math.random() * 2 + 1;
-  let rotation: number = Math.random() * 60 + 30;
-  let opacity: number = 1;
+  let rotation: number = $state(Math.random() * 60 + 30);
+  let opacity: number = $state(1);
   let animationFrame: number;
 
   function animate() {
@@ -52,12 +52,12 @@
     cancelAnimationFrame(animationFrame);
   });
 
-  $: style = `
+  let style = $derived(`
     position: absolute;
     font-size: 24px;
     transform: translate(${pos.x}px, ${pos.y}px) rotate(${rotation}deg);
     opacity: ${opacity};
-    `;
+    `);
 </script>
 
 <div {style} bind:this={elem}>🍪</div>
