@@ -4,12 +4,18 @@
   import { Body } from "svelte-body";
   import FallingCookie from "$lib/components/falling-cookie.svelte";
 
-  let fallingCookies: { id: number }[] = [];
+  interface Props {
+    children?: import("svelte").Snippet;
+  }
+
+  let { children }: Props = $props();
+
+  let fallingCookies: { id: number }[] = $state([]);
   let nextCookieId: number = 0;
 
-  let height: number = 0;
-  let scroll: number = 0;
-  let dissipate: boolean = false;
+  let height: number = $state(0);
+  let scroll: number = $state(0);
+  let dissipate: boolean = $derived(scroll > height * 0.6);
 
   function popOutCookie() {
     fallingCookies = [...fallingCookies, { id: nextCookieId++ }];
@@ -18,8 +24,6 @@
   function removeCookie(id: number) {
     fallingCookies = fallingCookies.filter((cookie) => cookie.id != id);
   }
-
-  $: dissipate = scroll > height * 0.6;
 </script>
 
 <svelte:window bind:innerHeight={height} bind:scrollY={scroll} />
@@ -52,7 +56,7 @@
 </div>
 
 <main id="top">
-  <slot />
+  {@render children?.()}
 </main>
 <hr />
 <footer>
@@ -63,7 +67,7 @@
   </p>
   <p>
     This website doesn't use cookies, so take as many as you want
-    <button id="gettingacookie" on:click={popOutCookie}>
+    <button id="gettingacookie" onclick={popOutCookie}>
       🍪
       {#each fallingCookies as cookie (cookie.id)}
         <FallingCookie on:remove={() => removeCookie(cookie.id)} />
